refactor: reorganize statistics dto class'es package to dto.repository package
