+++ src/main/java/com/example/spinlog/statistics/dto/repository/MemoDto.java
@@ -1,4 +1,4 @@
-package com.example.spinlog.statistics.dto;
+package com.example.spinlog.statistics.dto.repository;
 
 import lombok.*;
