+++ src/test/java/com/example/spinlog/integration/GenderStatisticsCacheUserWriteIntegrationTest.java
@@ -6,9 +6,9 @@ import com.example.spinlog.article.repository.ArticleRepository;
 import com.example.spinlog.global.security.customFilter.OAuth2ResponseImpl;
 import com.example.spinlog.global.security.oauth2.user.CustomOAuth2User;
 import com.example.spinlog.integration.init.RepositoryDataSetupService;
-import com.example.spinlog.statistics.dto.GenderDailyAmountSumDto;
-import com.example.spinlog.statistics.dto.GenderEmotionAmountAverageDto;
-import com.example.spinlog.statistics.dto.GenderSatisfactionAverageDto;
+import com.example.spinlog.statistics.dto.repository.GenderDailyAmountSumDto;
+import com.example.spinlog.statistics.dto.repository.GenderEmotionAmountAverageDto;
+import com.example.spinlog.statistics.dto.repository.GenderSatisfactionAverageDto;
 import com.example.spinlog.integration.init.GenderStatisticsCacheSetupService;
 import com.example.spinlog.statistics.service.cache.GenderStatisticsCacheFallbackService;
 import com.example.spinlog.user.dto.request.UpdateUserRequestDto;
