+++ src/test/java/com/example/spinlog/statistics/repository/GenderStatisticsRepositoryTest.java
@@ -4,10 +4,10 @@ import com.example.spinlog.article.entity.Article;
 import com.example.spinlog.article.entity.Emotion;
 import com.example.spinlog.article.entity.RegisterType;
 import com.example.spinlog.article.repository.ArticleRepository;
-import com.example.spinlog.statistics.dto.GenderDailyAmountSumDto;
-import com.example.spinlog.statistics.dto.GenderEmotionAmountAverageDto;
-import com.example.spinlog.statistics.dto.GenderSatisfactionAverageDto;
-import com.example.spinlog.statistics.dto.MemoDto;
+import com.example.spinlog.statistics.dto.repository.GenderDailyAmountSumDto;
+import com.example.spinlog.statistics.dto.repository.GenderEmotionAmountAverageDto;
+import com.example.spinlog.statistics.dto.repository.GenderSatisfactionAverageDto;
+import com.example.spinlog.statistics.dto.repository.MemoDto;
 import com.example.spinlog.user.entity.Gender;
 import com.example.spinlog.user.entity.Mbti;
 import com.example.spinlog.user.entity.User;
